Integrate AOCL code

+++ src/matrix_util.h
@@ -5,5 +5,5 @@
 bool EqualMatrices(const Matrix& a, const Matrix& b);
 void printMatrix(Matrix& a);
 void fillMatrixRandomly(Matrix& a);
-void fillMatrixRandomly(double* matrix, int rows, int columns);
+void fillMatrixRandomly(double* matrix, size_t rows, size_t columns);
 Matrix multiplyDecomposed(const Matrix& a, const Matrix& b, const Matrix& u, const Matrix& v, const Matrix& w);
